Make checking for strings in vega canvases more robust
